ai-engine/requirements.txt
mediapipe
opencv-python
numpy
python-socketio[asyncio_client]
aiohttp